bpf,tests: Change makefile cache to rebuild on header changes
[ upstream commit 49919dae73d8e61f081a10c61299f1bcf8210b81 ]

During BPF test development, it is common to make changes in
dependencies of the test C file and then re-test. However, because
we don't explicitly list header files and don't use them in the
traditional way of only containing forward declarations and using
other C files for the contents, the makefile doesn't understand it
needs to rebuild.

This commit makes clang emit a dependency file for each C file, this
dependency file is then fed into `make` so it is aware of all files
that contribute to the test C file, this makes it so any changes to
any dependent file including .h files will invalidate the `make` cache.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>
Signed-off-by: Maxim Mikityanskiy <maxim@isovalent.com>