make: add "run-builder" target
[ upstream commit b257fd50287ea6fa4cef41eec3d83c5f5a34c4c9 ]

Add make target for running bash in the cilium builder container.  This
provides quick access for the correct build environment, with the correct
go version for example. If backported to releases, this makes it easy to
to work with go vendoring with different releases, which each require a
different Go version to be able to run 'go mod' commands.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>
Signed-off-by: Maxim Mikityanskiy <maxim@isovalent.com>